Add files via upload
This is how I plan to download all the webms, currently it's only downloading the second row of the YLYL table, due to the fact I can't code and I hit many roadblocks along the way. However, I'm planning to fix that and optimize the code altogether in (hopefully) the next commit.